fix: tile border resize handles sometimes inaccessible on first render
